+++ app/components/UploadForm.tsx
@@ -4,6 +4,7 @@ import { useState } from 'react';
 import Image from 'next/image';
 import Modal from './Modal';
 import JumpBar from './JumpBar';
+import { getServiceYear } from '@/app/utils/serviceYear';
 
 interface FormData {
   userName: string;
@@ -15,6 +16,7 @@ interface FormData {
 
 const UploadForm = () => {
   const districtNumbers = Array.from({ length: 17 }, (_, i) => i + 1).filter(num => num !== 10);
+  const serviceYear = getServiceYear();
 
   const [formData, setFormData] = useState<FormData>({
     userName: '',
@@ -252,8 +254,8 @@ const UploadForm = () => {
           />
         </div>
 
-        <h1 className="text-3xl font-bold mb-8 text-center text-blue-800 uppercase">* 2025-2026 Report Submission Portal *</h1>
-        <p className="text-center text-black text-xl">Please fill out the form below to submit your reports for the 2025-2026 service year.</p>
+        <h1 className="text-3xl font-bold mb-8 text-center text-blue-800 uppercase">* {serviceYear} Report Submission Portal *</h1>
+        <p className="text-center text-black text-xl">Please fill out the form below to submit your reports for the {serviceYear} service year.</p>
         <p className="text-center text-black text-xl italic font-bold  mb-8">Send any issues or questions to <a href="mailto:adjutant@floridasons.org" className="text-blue-500">adjutant@floridasons.org</a></p>
         {/* User Information Form */}
         <div className="bg-white shadow-md rounded-lg p-6 mb-8">